Move the cookie storage adapter into `shared`
