+++ src/lib/features/listening-history/model/cookie-name.ts
@@ -1,0 +1,1 @@
+export const cookieName = 'listening-history';
