+++ src/lib/shared/lib/cookie-persistence/cookie-storage-adapter.ts
@@ -1,20 +1,25 @@
 import Cookies from 'js-cookie';
 import { browser } from '$app/env';
-
-export const cookieName = 'listening-history';
+import type { ItemStorage } from 'svelte-persistent-writable';
 
 /**
  * The adapter to store data in cookies on the client.
  *
  * On the server side, the operations are no-ops.
  */
-export const cookieStorageAdapter = {
+export class CookieStorageAdapter implements ItemStorage<number[]> {
+  private cookieName: string;
+
+  constructor(cookieName: string) {
+    this.cookieName = cookieName;
+  }
+
   get(): { value: number[] } | null {
     if (!browser) {
       return null;
     }
 
-    const cookieValue = Cookies.get(cookieName);
+    const cookieValue = Cookies.get(this.cookieName);
     if (cookieValue === undefined) {
       return null;
     } else {
@@ -24,23 +29,25 @@ export const cookieStorageAdapter = {
         return { value: [] };
       }
     }
-  },
+  }
+
   set(ids: number[]) {
     if (!browser) {
       return;
     }
 
-    Cookies.set(cookieName, JSON.stringify(ids), {
+    Cookies.set(this.cookieName, JSON.stringify(ids), {
       expires: 99999,
       secure: true,
       sameSite: 'Strict',
     });
-  },
+  }
+
   remove() {
     if (!browser) {
       return;
     }
 
-    Cookies.remove(cookieName);
-  },
-};
+    Cookies.remove(this.cookieName);
+  }
+}
