+++ src/lib/shared/lib/index.ts
@@ -4,3 +4,4 @@ export type {
   CheckStatusRequest,
   QueryDownloadsRequest,
 } from './service-worker/message';
+export { CookieStorageAdapter } from './cookie-persistence/cookie-storage-adapter';
